+++ src/main/java/cart/controller/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.controller.dto;
 
 public class ErrorResponse {
     private final String message;
