+++ src/main/java/cart/controller/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.controller.dto;
 
 import cart.entity.Member;
