+++ src/main/java/cart/controller/dto/ProductResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.controller.dto;
 
 import cart.entity.Product;
